+++ changelog.md
@@ -23,3 +23,4 @@
 7. Correctly support Python Sequence for `utils.build_texture_from_PyTinyrenderer` as texture.
 8. `quaternion` function to create an orientation from axis and angle, and `quaternion_mul` to composite quaternion.
 9. `rotation_matrix` function to create a rotation matrix from axis and angle. Also allows `Scene` to set object orientation directly using rotation matrix.
+10. Move `Renderer.merge_objects` into `geometry.py`, and expose in `__init__.py`.